// apps/reatom/src/shared/api/schema.ts
export interface components {
  schemas: {
// ...
    GetTracksRequestPayload: {
      /**
       * @description Page number for pagination (starting from 1)
       * @default 1
       */
      pageNumber: number
// ...
      tagsIds?: string[]
// ...
    }
// ...
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}
// ...
export type SchemaGetPlaylistsRequestPayload = components['schemas']['GetPlaylistsRequestPayload']
// ...
export enum PathsPlaylistsGetParametersQuerySortBy {
  addedAt = 'addedAt',
  likesCount = 'likesCount',
}
export enum PathsPlaylistsGetParametersQuerySortDirection {
  asc = 'asc',
  desc = 'desc',
}